Deprecate `"core/geometry"` folder
- deprecate notice for "core/geometry" folder (will be removed after v3.4)
- move entire parsing logic within "core/utils/geo"
