+++ src/app/g3w-ol/controls/querybypolygoncontrol.js
@@ -2,7 +2,7 @@ import {SPATIALMETHODS, VM} from '../constants';
 const {merge} = require('../utils');
 const InteractionControl = require('./interactioncontrol');
 const PickCoordinatesInteraction = require('../interactions/pickcoordinatesinteraction');
-const {getAllPolygonGeometryTypes} = require('core/geometry/geometry');
+const {getAllPolygonGeometryTypes} = require('core/utils/geo').Geometry;
 const VALIDGEOMETRIES = getAllPolygonGeometryTypes();
 
 const QueryByPolygonControl = function(options={}) {
