+++ src/app/g3w-ol/utils/utils.js
@@ -1,5 +1,5 @@
 import ApplicationState from "../../core/applicationstate";
-const Geometry = require('core/geometry/geometry');
+const { Geometry } = require('core/utils/geo');
 
 const INCHES_PER_UNIT = {
   m: 39.37, //
